requirements.txt
jupyter
jupyterlab
pandas
todoist-python
python-dotenv
click
